+++ ethereum/backend.go
@@ -54,6 +54,7 @@ func Backend(conf *logical.BackendConfig) (*PluginBackend, error) {
 			ConfigPaths(&b),
 			WalletPaths(&b),
 			ERC20Paths(&b),
+			PlasmaPaths(&b),
 			AccountPaths(&b),
 			ExportPaths(&b),
 		),